mem: Check if buffer is not empty before writing to channel
If a call to Close was made on the mem.MessageWriter without ever
calling Write, it would still write a msg.Message to the channel with an
empty Body. This caused any Receiver to get a bogus message.

Instead, this adds a check to make sure the buffer is not empty before
creating a msg.Message and writing it to the channel.